requirements.txt
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
asttokens==3.0.0
Authlib==1.3.1
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
comm==0.2.2
cryptography==44.0.1
debugpy==1.8.12
decorator==5.1.1
executing==2.2.0
fastapi==0.115.8
grpcio==1.70.0
grpcio-health-checking==1.70.0
grpcio-tools==1.70.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==26.2.1
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.45.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
uvicorn==0.34.0
validators==0.34.0
wcwidth==0.2.13
weaviate-client==4.10.4
